feat: cleanup and refactor code
